requirements.txt
django>=1.7.0
djangorestframework>=3.0.0